changed OdysseusNet post processor "DiscardedReplicates"
Post processor to insert a {@link RecoveryTimeCalculatorAO}(!!!) and a sender for each {@link ReplicationMergeAO}. The sender will be inserted for the output port 1 that is not used normally. All discarded replicates are sent to port 1. The sender writes the data in a CSV file (one file per merger). The only argument for this post processor is the path to the CSV files. The names of the files are determined by the {@link ReplicationMergeAO} (name and hashCode).

Used sender settings:
- transport handler "file"
- data hgandler "tuple"
- wrapper "GenericPush"
- protocol handler "csv"
- "createDir" "true"
- "filename", path_from_user + "/" + merger.getName() + merger.hashCode() + ".csv"
- "append", "true"
- "writeMetaData", "true"

git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@27797 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: e2956b47c93db010230db8e9f561780279bb93ec