Added automatic predicate-changing
If you query over encrypted data, you may want to encrypt predicates as
well. they will be crypted, set and changed (e.g. if the cryptor uses a
new key) automatically.
Because of this, a rule has to set SecurityPredicates, which can be
crypted, to IHasPredicate and IHasPredicates.
Also the regularExpression will be crypted, so it has to be enabled to
change the SDFExpression.

Former-commit-id: 0d28c56e6e9c9fee1b35a16bde8ce0d6c303c2c5