ODY-1403 (Hopefully) solved a conceptional error
When doing trajectory prediction (with estimation / prefiltering), you
have to 1. predict the points of the center object, then for each
predicted point making the same thing as if you do a prediction to only
one point: estimate possible results (prefiltering) and then predicting
the other object. Then you can do a normal radius query.

Now, the prediction operator can be used for both and the extra
operators for the case that you want to predict a trajectory are no
longer needed. Way more elegant if you ask me.

Former-commit-id: 34b0f058161225386387f5e131307cb8e9577822