ODY-1520 Worked on some bugs which came from the constraints-handling
(1) Simple copy instructions from a map operator resulted in a
transformation to a generic temporal type, even though the original
temporal type could be simply copied / left in the result. Now, this is
done by checking if it is a simple copy instruction (no real expression)
and if so, simply using a normal instead of a temporal relational
expression (TemporalRelationalMapPO.java).

(2) When using keepInput or remove some elements in a map operator, the
INIT processor TSetTemporalConstraintsOnMapAORule reduced the schema to
only the expressions. Example: You have 1 expression and set keepInput
to true. The input schema has 9 attributes. The output schema therefore
has 10 attributes. But the INIT processor removed everything except the
expression attribute. Now, the loop is doing its work the other way
around. Instead of looping through the expressions, it loops through the
correct output schema and if it finds an expression which needs to have
an constraint added, it does so. But is keeps the rest. That's better,
not loosing all this important information.

(3) The TTemporalSelectAORule did not fire, even if it should. Too sad.
This was because the attributes in the predicate did not change by the
INIT processor which adds the temporal constraints. And no constraints,
no TemporalSelect. Now it takes a look at the input schema and looks if
this has a temporal attribute which is used by a predicate. Works for
now, but I think I have to check the qual name or so to get the cases
where the user writes predicates with the source name (source.attribute
> 1) or so.

Clean code with comments and so on comes soon.

Former-commit-id: 0d3c637227602efe3be9301df8e352e9b799f6f7