ODY-1492 Clean in all cases when processing outdating element to avoid
old tuples being there forever if for that trigger point in time no
evaluation is necessary (due to the boolean expression at the beginning
of that method.

What I'm thinking about: do we remove elements that are maybe needed for
other triggers with the same start timestamp? I don't think so because
in that case the other remove statement if it is evaluated would have
the same problem, wouldn't it?

Former-commit-id: ad2adf8163d0478dfc2b617560060602547de781