ODY-1605 Use the less granular time unit for the merge function.
This is necessary because it could be possible that the less granular
temporal type does not have the values in-between the points in time.
This is for example the case if a GenericTemporalType is used. But it
would be necessary to have those to work with the more granular time
unit. Therefore use the less granular one.

Former-commit-id: 04ee17137b0a533b94771b2859d4ac7ced73d7f0