104ProtocolHandler: Avoid multithreading
The minimal solution to avoid multithreading and to not touch the
library is to implement an own handler for the APDUs
(OdysseusAPDUHandler). With that, I have no more problems of not the
same amount of tuples in different runs in the result set. It also makes
the mapping of timestamp to APDU easier.

Former-commit-id: b93e09116a9b201bea60a13739915f75dc9b30da